requirements.txt
opencv_python_headless==4.2.0.32
torch==1.3.1
torchvision==0.4.2
numpy==1.16.2
Flask==1.1.1
Pillow==7.0.0
gunicorn==20.0.4

